Merged PR 144: Fix performance bug, exclude receipt creation in measurement, add switch to control detail time measurement.
Do not update dep when no-overmatching is true.
Do not measure detailed time consumption by default (--detailtime to turn it on)
Wrap txpreplay cache with simpletrylock so that process is not blocked during peeking txpreplay
Add more pause points in cacheEvictionLoop2 to avoid txpreplay cache contention with process
Exclude receipt creation in measurement.